pyproject.toml
[project]
name = "moodle-claude"
version = "0.1.0"
description = "MCP server for creating Moodle courses from Claude chat conversations"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "aiosqlite>=0.21.0",
    "markdown>=3.8.2",
    "mcp>=1.12.2",
    "pygments>=2.19.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "responses>=0.25.7",
]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
]
